Move dedicated algorithms to their respective layers
