+++ src/backend/tfjs_backend.ts
@@ -16,9 +16,9 @@
 import * as tfc from '@tensorflow/tfjs-core';
 import {DataType, dispose, onesLike as coreOnesLike, Scalar, scalar, Tensor, Tensor1D, tensor1d, Tensor2D, tensor2d, Tensor3D, Tensor4D, tidy, util, variableGrads, where, zerosLike as coreZerosLike} from '@tensorflow/tfjs-core';
 
-import {checkDataFormat, checkPaddingMode, checkPoolMode, DataFormat, nameScope as commonNameScope, PaddingMode, PoolMode} from '../common';
+import {checkDataFormat, DataFormat, nameScope as commonNameScope} from '../common';
 import {NotImplementedError, ValueError} from '../errors';
-import {RnnStepFunction, Shape, SymbolicTensor} from '../types';
+import {Shape, SymbolicTensor} from '../types';
 import * as math_utils from '../utils/math_utils';
 import {LayerVariable} from '../variables';
 import {epsilon as common_epsilon} from './common';
@@ -176,83 +176,6 @@ export function expandDims(x: Tensor, axis = -1): Tensor {
   }
   outShape.splice(axis, 0, 1);
   return reshape(x, outShape);
-}
-
-
-/**
- * Pads the middle dimension of a 3D tensor.
- *
- * @param x Input `Tensor` to be padded.
- * @param padding `Array` of 2 integers, how many zeros to add at the start and
- *   end of the middle dimension (i.e., dimension 1).
- * @return A padded 3D `Tensor`.
- */
-export function temporalPadding(x: Tensor, padding?: [number, number]): Tensor {
-  if (ndim(x) !== 3) {
-    throw new ValueError(
-        `temporalPadding expects input tensor to be 3-D, but received a ` +
-        `${ndim(x)}-D tensor.`);
-  }
-
-  if (padding == null) {
-    padding = [1, 1];
-  }
-  if (padding.length !== 2) {
-    throw new ValueError(
-        `temporalPadding expects input padding pattern to be a length-2 ` +
-        `array, but received a length-${padding.length} array.`);
-  }
-
-  const pattern: Array<[number, number]> = [[0, 0], padding, [0, 0]];
-  return tfc.pad(x, pattern);
-}
-
-/**
- * Pads the 2nd and 3rd dimensions of a 4D tensor.
- *
- * @param x Input `Tensor` to be padded.
- * @param padding `Array` of two `Array`s, each of which is an `Array` of two
- *   integers. The amount of padding at the beginning and end of the 2nd and 3rd
- *   dimensions, respectively.
- * @param dataFormat 'channelsLast' (default) or 'channelsFirst'.
- * @return Padded 4D `Tensor`.
- */
-export function spatial2dPadding(
-    x: Tensor, padding?: [[number, number], [number, number]],
-    dataFormat?: DataFormat): Tensor {
-  if (ndim(x) !== 4) {
-    throw new ValueError(
-        `temporalPadding expects input tensor to be 4-D, but received a ` +
-        `${ndim(x)}-D tensor.`);
-  }
-
-  if (padding == null) {
-    padding = [[1, 1], [1, 1]];
-  }
-  if (padding.length !== 2 || padding[0].length !== 2 ||
-      padding[1].length !== 2) {
-    throw new ValueError(
-        'spatial2dPadding expects `padding` to be an Array of two Arrays, ' +
-        'each of which is an Array of two integers.');
-  }
-
-  if (dataFormat == null) {
-    dataFormat = imageDataFormat();
-  }
-  if (dataFormat !== 'channelsLast' && dataFormat !== 'channelsFirst') {
-    throw new ValueError(
-        `Unknown data format: ${dataFormat}. ` +
-        `Supported data formats are 'channelsLast' and 'channelsFirst.`);
-  }
-
-  let pattern: Array<[number, number]>;
-  if (dataFormat === 'channelsFirst') {
-    pattern = [[0, 0], [0, 0], padding[0], padding[1]];
-  } else {
-    pattern = [[0, 0], padding[0], padding[1], [0, 0]];
-  }
-
-  return tfc.pad(x, pattern);
 }
 
 /**
@@ -1110,251 +1033,6 @@ export function l2Normalize(x: Tensor, axis?: number): Tensor {
 }
 
 /**
- * Transpose and cast the input before the conv2d.
- * @param x Input image tensor.
- * @param dataFormat
- */
-function preprocessConv2DInput(x: Tensor, dataFormat: DataFormat): Tensor {
-  // TODO(cais): Cast type to float32 if not.
-  checkDataFormat(dataFormat);
-  if (dataFormat === 'channelsFirst') {
-    return tfc.transpose(x, [0, 2, 3, 1]);  // NCHW -> NHWC.
-  } else {
-    return x;
-  }
-}
-
-/**
- * 1D-convolution with bias added.
- *
- * Porting Note: This function does not exist in the Python Keras backend.
- *   It is exactly the same as `conv2d`, except the added `bias`.
- *
- * @param x Input tensor, rank-3, of shape `[batchSize, width, inChannels]`.
- * @param kernel Kernel, rank-3, of shape `[filterWidth, inDepth, outDepth]`.
- * @param bias Bias, rank-3, of shape `[outDepth]`.
- * @param strides
- * @param padding Padding mode.
- * @param dataFormat Data format.
- * @param dilationRate
- * @returns The result of the 1D convolution.
- * @throws ValueError, if `x`, `kernel` or `bias` is not of the correct rank.
- */
-export function conv1dWithBias(
-    x: Tensor, kernel: Tensor, bias: Tensor, strides = 1, padding = 'valid',
-    dataFormat?: DataFormat, dilationRate = 1): Tensor {
-  if (dataFormat == null) {
-    dataFormat = imageDataFormat();
-  }
-  checkDataFormat(dataFormat);
-
-  // Check the ranks of x, kernel and bias.
-  if (x.shape.length !== 3) {
-    throw new ValueError(
-        `The input of a conv1dWithBias operation should be 3, but is ` +
-        `${x.shape.length} instead.`);
-  }
-  if (kernel.shape.length !== 3) {
-    throw new ValueError(
-        `The kernel for a conv1dWithBias operation should be 3, but is ` +
-        `${kernel.shape.length} instead`);
-  }
-  if (bias != null && bias.shape.length !== 1) {
-    throw new ValueError(
-        `The bias for a conv1dWithBias operation should be 1, but is ` +
-        `${kernel.shape.length} instead`);
-  }
-
-  // TODO(cais): Support CAUSAL padding mode.
-
-  if (dataFormat === 'channelsFirst') {
-    x = tfc.transpose(x, [0, 2, 1]);  // NCW -> NWC.
-  }
-  if (padding === 'causal') {
-    throw new NotImplementedError(
-        'The support for CAUSAL padding mode in conv1dWithBias is not ' +
-        'implemented yet.');
-  }
-  let y: Tensor = tfc.conv1d(
-      x as Tensor2D | Tensor3D, kernel as Tensor3D, strides,
-      padding === 'same' ? 'same' : 'valid', 'NWC', dilationRate);
-  if (bias != null) {
-    y = biasAdd(y, bias);
-  }
-  return y;
-}
-
-/**
- * 1D-convolution.
- *
- * @param x Input tensor, rank-3, of shape `[batchSize, width, inChannels]`.
- * @param kernel Kernel, rank-3, of shape `[filterWidth, inDepth, outDepth]`.s
- * @param strides
- * @param padding Padding mode.
- * @param dataFormat Data format.
- * @param dilationRate
- * @returns The result of the 1D convolution.
- * @throws ValueError, if `x`, `kernel` or `bias` is not of the correct rank.
- */
-export function conv1d(
-    x: Tensor, kernel: Tensor, strides = 1, padding = 'valid',
-    dataFormat?: DataFormat, dilationRate = 1): Tensor {
-  checkDataFormat(dataFormat);
-  return conv1dWithBias(
-      x, kernel, null, strides, padding, dataFormat, dilationRate);
-}
-
-/**
- * 2D Convolution
- * @param x
- * @param kernel kernel of the convolution.
- * @param strides strides array.
- * @param padding padding mode. Default to 'valid'.
- * @param dataFormat data format. Defaults to 'channelsLast'.
- * @param dilationRate dilation rate array.
- * @returns Result of the 2D pooling.
- */
-export function conv2d(
-    x: Tensor, kernel: Tensor, strides = [1, 1], padding = 'valid',
-    dataFormat?: DataFormat, dilationRate?: [number, number]): Tensor {
-  checkDataFormat(dataFormat);
-  return conv2dWithBias(
-      x, kernel, null, strides, padding, dataFormat, dilationRate);
-}
-
-/**
- * 2D Convolution with an added bias.
- * Note: This function does not exist in the Python Keras Backend. This function
- * is exactly the same as `conv2d`, except the added `bias`.
- */
-export function conv2dWithBias(
-    x: Tensor, kernel: Tensor, bias: Tensor, strides = [1, 1],
-    padding = 'valid', dataFormat?: DataFormat,
-    dilationRate?: [number, number]): Tensor {
-  if (dataFormat == null) {
-    dataFormat = imageDataFormat();
-  }
-  checkDataFormat(dataFormat);
-  if (ndim(x) !== 3 && ndim(x) !== 4) {
-    throw new ValueError(
-        `conv2dWithBias expects input to be of rank 3 or 4, but received ` +
-        `${ndim(x)}.`);
-  }
-  if (ndim(kernel) !== 3 && ndim(kernel) !== 4) {
-    throw new ValueError(
-        `conv2dWithBias expects kernel to be of rank 3 or 4, but received ` +
-        `${ndim(x)}.`);
-  }
-  let y = preprocessConv2DInput(x, dataFormat);
-  if (padding === 'causal') {
-    throw new NotImplementedError(
-        'The support for CAUSAL padding mode in conv1dWithBias is not ' +
-        'implemented yet.');
-  }
-  y = tfc.conv2d(
-      y as Tensor3D | Tensor4D, kernel as Tensor4D, strides as [number, number],
-      padding === 'same' ? 'same' : 'valid', 'NHWC', dilationRate);
-  if (bias != null) {
-    y = biasAdd(y, bias as Tensor1D);
-  }
-  if (dataFormat === 'channelsFirst') {
-    y = tfc.transpose(y, [0, 3, 1, 2]);
-  }
-  return y;
-}
-
-/**
- * 2D convolution with separable filters.
- * @param x Input tensor.
- * @param depthwiseKernel Convolution kernel for depthwise convolution.
- * @param strides Strides (Array of two integers).
- * @param padding Padding model.
- * @param dataFormat Data format.
- * @param dilationRate Array of two integers, dilation rates for the separable
- *   convolution.
- * @returns Output tensor.
- * @throws ValueError If depthwiseKernel is not a 4D array.
- */
-export function depthwiseConv2d(
-    x: Tensor, depthwiseKernel: Tensor, strides: [number, number] = [1, 1],
-    padding = 'valid', dataFormat?: DataFormat,
-    dilationRate?: [number, number]): Tensor {
-  if (dataFormat == null) {
-    dataFormat = imageDataFormat();
-  }
-  checkDataFormat(dataFormat);
-  let y = preprocessConv2DInput(x, dataFormat);
-  if (ndim(x) !== 4) {
-    throw new ValueError(
-        `Input for depthwiseConv2d is required to be 4-D, but is instead ` +
-        `${ndim(x)}-D`);
-  }
-  if (ndim(depthwiseKernel) !== 4) {
-    throw new ValueError(
-        `depthwiseKernel is required to be 4-D, but is instead ` +
-        `${ndim(depthwiseKernel)}-D`);
-  }
-  y = tfc.depthwiseConv2d(
-      y as Tensor4D, depthwiseKernel as Tensor4D, strides,
-      padding === 'same' ? 'same' : 'valid', 'NHWC', dilationRate);
-  if (dataFormat === 'channelsFirst') {
-    y = tfc.transpose(y, [0, 3, 1, 2]);
-  }
-  return y;
-}
-
-/**
- * 2D pooling.
- * @param x
- * @param poolSize
- * @param stridesdes strides. Defaults to [1, 1].
- * @param padding padding. Defaults to 'valid'.
- * @param dataFormat data format. Defaults to 'channelsLast'.
- * @param poolMode Mode of pooling. Defaults to 'max'.
- * @returns Result of the 2D pooling.
- */
-export function pool2d(
-    x: Tensor, poolSize: [number, number], strides?: [number, number],
-    padding?: PaddingMode, dataFormat?: DataFormat,
-    poolMode?: PoolMode): Tensor {
-  checkDataFormat(dataFormat);
-  checkPoolMode(poolMode);
-  checkPaddingMode(padding);
-  if (strides == null) {
-    strides = [1, 1];
-  }
-  if (padding == null) {
-    padding = 'valid';
-  }
-  if (dataFormat == null) {
-    dataFormat = imageDataFormat();
-  }
-  if (poolMode == null) {
-    poolMode = 'max';
-  }
-
-  // TODO(cais): Remove the preprocessing step once deeplearn.js supports
-  // dataFormat as an input argument.
-  x = preprocessConv2DInput(x, dataFormat);  // x is NHWC after preprocessing.
-  let y: Tensor;
-  const paddingString = (padding === 'same') ? 'same' : 'valid';
-  if (poolMode === 'max') {
-    // TODO(cais): Rank check?
-    y = tfc.maxPool(x as Tensor4D, poolSize, strides, paddingString);
-  } else {  // 'avg'
-    // TODO(cais): Check the dtype and rank of x and give clear error message
-    //   if those are incorrect.
-    y = tfc.avgPool(
-        // TODO(cais): Rank check?
-        x as Tensor3D | Tensor4D, poolSize, strides, paddingString);
-  }
-  if (dataFormat === 'channelsFirst') {
-    y = tfc.transpose(y, [0, 3, 1, 2]);  // NHWC -> NCHW.
-  }
-  return y;
-}
-
-/**
  * Replacement for Keras's "with name_scope" construct.
  *
  * @param name The name to use for this name scope.
@@ -1514,116 +1192,6 @@ export function inTrainPhase<T>(x: () => T, alt: () => T, training = false): T {
 /**
  * Control flow.
  */
-
-/**
- * Iterates over the time dimension of a tensor.
- *
- * @param stepFunction RNN step function.
- *   Parameters:
- *     inputs: tensor with shape `[samples, ...]` (no time dimension),
- *       representing input for the batch of samples at a certain time step.
- *     states: an Array of tensors.
- *   Returns:
- *     outputs: tensor with shape `[samples, outputDim]` (no time dimension).
- *     newStates: list of tensors, same length and shapes as `states`. The first
- *       state in the list must be the output tensor at the previous timestep.
- * @param inputs Tensor of temporal data of shape `[samples, time, ...]` (at
- *   least 3D).
- * @param initialStates Tensor with shape `[samples, outputDim]` (no time
- *   dimension), containing the initial values of the states used in the step
- *   function.
- * @param goBackwards If `true`, do the iteration over the time dimension in
- *   reverse order and return the reversed sequence.
- * @param mask Binary tensor with shape `[sample, time, 1]`, with a zero for
- *   every element that is masked.
- * @param constants An Array of constant values passed at each step.
- * @param unroll Whether to unroll the RNN or to use a symbolic loop. *Not*
- *   applicable to this imperative deeplearn.js backend. Its value is ignored.
- * @param inputLength Not relevant in this deeplearn.js backend.
- * @returns An Array: `[lastOutput, outputs, newStates]`.
- *   lastOutput: the lastest output of the RNN, of shape `[samples, ...]`.
- *   outputs: tensor with shape `[samples, time, ...]` where each entry
- *     `output[s, t]` is the output of the step function at time `t` for sample
- *     `s`.
- *   newStates: Array of tensors, latest states returned by the step function,
- *      of shape `(samples, ...)`.
- * @throws ValueError If input dimension is less than 3.
- */
-export function rnn(
-    stepFunction: RnnStepFunction, inputs: Tensor, initialStates: Tensor[],
-    goBackwards = false, mask?: Tensor, constants?: Tensor[], unroll = false,
-    inputLength?: number): [Tensor, Tensor, Tensor[]] {
-  const ndim = inputs.shape.length;
-  if (ndim < 3) {
-    throw new ValueError(`Input should be at least 3D, but is ${ndim}D.`);
-  }
-
-  // Transpose to time-major, i.e., from [batch, time, ...] to [time, batch,
-  // ...].
-  const axes = [1, 0].concat(math_utils.range(2, ndim));
-  inputs = tfc.transpose(inputs, axes);
-
-  if (mask != null) {
-    throw new NotImplementedError(
-        'The rnn() function of the deeplearn.js backend does not support ' +
-        'masking yet.');
-  }
-
-  if (constants != null) {
-    throw new NotImplementedError(
-        'The rnn() functoin of the deeplearn.js backend does not support ' +
-        'constants yet.');
-  }
-
-  // Porting Note: the unroll option is ignored by the imperative backend.
-  if (unroll) {
-    console.warn(
-        'Backend rnn(): the unroll = true option is not applicable to the ' +
-        'imperative deeplearn.js backend.');
-  }
-
-  if (goBackwards) {
-    inputs = tfc.reverse(inputs, 0);
-  }
-
-  // Porting Note: PyKeras with TensorFlow backend uses a symbolic loop
-  //   (tf.while_loop). But for the imperative deeplearn.js backend, we just use
-  //   the usual TypeScript control flow to iterate over the time steps in the
-  //   inputs.
-  // Porting Note: PyKeras patches a "_use_learning_phase" attribute to outputs.
-  //   This is not idiomatic in TypeScript. The info regarding whether we are
-  //   in a learning (i.e., training) phase for RNN is passed in a different
-  //   way.
-  //   TODO(cais): Determine in what exact way the learning phase information
-  //     will be passed.
-
-  let outputs: Tensor;
-  let lastOutput: Tensor;
-  let states = initialStates;
-  const timeSteps = inputs.shape[0];
-  for (let t = 0; t < timeSteps; ++t) {
-    let currentInput = sliceAlongFirstAxis(inputs, t, 1);
-    currentInput = reshape(currentInput, currentInput.shape.slice(1));
-    const stepOutputs = stepFunction(currentInput, states);
-    lastOutput = stepOutputs[0];
-    if (t === 0) {
-      outputs = lastOutput.reshape([1].concat(lastOutput.shape));
-    } else {
-      outputs = concatAlongFirstAxis(
-          outputs, lastOutput.reshape([1].concat(lastOutput.shape)));
-    }
-    // TODO(soergel): Call K.concatenate() to perform only one concatenation at
-    // the end, once the backend function is available.
-    states = stepOutputs[1];
-  }
-
-  return [
-    lastOutput,
-    tfc.transpose(
-        outputs, [1, 0].concat(math_utils.range(2, outputs.shape.length))),
-    states
-  ];
-}
 
 /**
  * Returns the gradients of `variables` w.r.t. the return value of `lossFn`.
