+++ src/backend/tfjs_backend_test.ts
@@ -14,9 +14,8 @@
 
 // tslint:disable:max-line-length
 import * as tfc from '@tensorflow/tfjs-core';
-import {DataType, Scalar, scalar, slice, Tensor, memory, tensor1d, tensor2d, tensor3d, tensor4d, Tensor4D, zeros} from '@tensorflow/tfjs-core';
+import {DataType, Scalar, scalar, Tensor, memory, tensor1d, tensor2d, tensor3d, tensor4d, zeros} from '@tensorflow/tfjs-core';
 
-import {DataFormat, PaddingMode, PoolMode} from '../common';
 import {SymbolicTensor} from '../types';
 import {LayerVariable} from '../variables';
 import {unique} from '../utils/generic_utils';
@@ -280,58 +279,6 @@ describeMathCPUAndGPU('expandDims', () => {
     const x = tensor2d([[10, 20], [30, 40]], [2, 2]);
     expectTensorsClose(
       K.expandDims(x, 0), tensor3d([[[10, 20], [30, 40]]], [1, 2, 2]));
-  });
-});
-
-describeMathCPUAndGPU('temporalPadding', () => {
-  it('default padding 1-1', () => {
-    const x = tensor3d([[[1, 2], [3, 4]], [[-1, -2], [-3, -4]]]);
-    const y = K.temporalPadding(x);
-    expectTensorsClose(y, tensor3d([
-      [[0, 0], [1, 2], [3, 4], [0, 0]],
-      [[0, 0], [-1, -2], [-3, -4], [0, 0]]
-    ]));
-  });
-  it('custom padding 2-2', () => {
-    const x = tensor3d([[[1, 2], [3, 4]], [[-1, -2], [-3, -4]]]);
-    const y = K.temporalPadding(x, [2, 2]);
-    expectTensorsClose(
-      y, tensor3d([
-        [[0, 0], [0, 0], [1, 2], [3, 4], [0, 0], [0, 0]],
-        [[0, 0], [0, 0], [-1, -2], [-3, -4], [0, 0], [0, 0]]
-      ]));
-  });
-});
-
-describeMathCPUAndGPU('spatial2dPadding', () => {
-  it('default padding 1-1-1-1', () => {
-    const x = tfc.ones([2, 3, 4, 3]);
-    const y = K.spatial2dPadding(x);
-
-    expect(y.shape).toEqual([2, 5, 6, 3]);
-    expectTensorsClose(slice(y, [0, 1, 1, 0], [2, 3, 4, 3]), x);
-    expectTensorsClose(
-      slice(y, [0, 0, 0, 0], [2, 1, 6, 3]), tfc.zeros([2, 1, 6, 3]));
-    expectTensorsClose(
-      slice(y, [0, 4, 0, 0], [2, 1, 6, 3]), tfc.zeros([2, 1, 6, 3]));
-    expectTensorsClose(
-      slice(y, [0, 0, 0, 0], [2, 5, 1, 3]), tfc.zeros([2, 5, 1, 3]));
-    expectTensorsClose(
-      slice(y, [0, 0, 5, 0], [2, 5, 1, 3]), tfc.zeros([2, 5, 1, 3]));
-  });
-
-  it('custom padding 2-2-3-0', () => {
-    const x = tfc.ones([2, 3, 4, 3]);
-    const y = K.spatial2dPadding(x, [[2, 2], [3, 0]]);
-    expect(y.shape).toEqual([2, 7, 7, 3]);
-
-    expectTensorsClose(slice(y, [0, 2, 3, 0], [2, 3, 4, 3]), x);
-    expectTensorsClose(
-      slice(y, [0, 0, 0, 0], [2, 2, 7, 3]), tfc.zeros([2, 2, 7, 3]));
-    expectTensorsClose(
-      slice(y, [0, 5, 0, 0], [2, 2, 7, 3]), tfc.zeros([2, 2, 7, 3]));
-    expectTensorsClose(
-      slice(y, [0, 0, 0, 0], [2, 7, 3, 3]), tfc.zeros([2, 7, 3, 3]));
   });
 });
 
@@ -1420,376 +1367,6 @@ describeMathCPUAndGPU('softsign', () => {
   });
 });
 
-describeMathCPUAndGPU('conv1dWithBias', () => {
-  const xLength4Data = [10, 20, 40, 80];
-  const kernelLength2Data = [1, -1];
-  const biasScalarData = 2.2;
-  // In the basic case, this convolves [10, 20, 40, 80] with the kernel [1, -1],
-  // producing [-10, -20, -40], and adds the bias 2.2, producing
-  // [-7.8, -17.8, -37.7].  The test is reproduced for either 1 or 2 output
-  // channels, and several reasonable data formats.
-
-  const outChannelsArray = [1, 2];
-  const dataFormats: DataFormat[] =
-    [undefined, 'channelsFirst', 'channelsLast'];
-  const paddingModes: PaddingMode[] = [undefined, 'same', 'valid'];
-  const stride = 1;
-
-  for (const outChannels of outChannelsArray) {
-    for (const dataFormat of dataFormats) {
-      for (const paddingMode of paddingModes) {
-        const testTitle = `outChannels=${outChannels}, stride=${stride}, ` +
-          `${paddingMode}, ${dataFormat}`;
-        it(testTitle, () => {
-          let x: Tensor = tensor3d(xLength4Data, [1, 4, 1]);
-          if (dataFormat === 'channelsFirst') {
-            x = tfc.transpose(x, [0, 2, 1]);  // NWC -> NCW.
-          }
-
-          let kernelData: number[] = [];
-          let biasData: number[] = [];
-          for (let i = 0; i < outChannels; ++i) {
-            kernelData = kernelData.concat(kernelLength2Data);
-            biasData = biasData.concat([biasScalarData + i]);
-          }
-          const kernel =
-            tfc.transpose(tensor3d(kernelData, [1, outChannels, 2]), [2, 0, 1]);
-          const bias = tensor1d(biasData);
-
-          const y = K.conv1dWithBias(
-            x, kernel, bias, stride, paddingMode, dataFormat);
-
-          let yExpectedShape: [number, number, number];
-          let yExpectedData: number[];
-          if (paddingMode === 'valid' || paddingMode === undefined) {
-            if (outChannels === 1) {
-              yExpectedShape = [1, 3, 1];
-              yExpectedData = [-7.8, -17.8, -37.8];
-            } else if (outChannels === 2) {
-              yExpectedShape = [1, 3, 2];
-              yExpectedData = [-7.8, -6.8, -17.8, -16.8, -37.8, -36.8];
-            }
-          } else if (paddingMode === 'same') {
-            if (outChannels === 1) {
-              yExpectedShape = [1, 4, 1];
-              yExpectedData = [-7.8, -17.8, -37.8, 82.2];
-            } else if (outChannels === 2) {
-              yExpectedShape = [1, 4, 2];
-              yExpectedData =
-                [-7.8, -6.8, -17.8, -16.8, -37.8, -36.8, 82.2, 83.2];
-            }
-          }
-          expectTensorsClose(y, tensor3d(yExpectedData, yExpectedShape));
-        });
-      }
-    }
-  }
-});
-
-describeMathCPUAndGPU('conv1d', () => {
-  const xLength4Data = [10, 20, 40, 80];
-  const kernelLength2Data = [1, -1];
-
-  const stride = 2;
-  const outChannels = 2;
-  const dataFormat = 'channelsLast';
-  const paddingMode = 'valid';
-  const testTitle = `outChannels=${outChannels}, stride=${stride}, ` +
-    `${paddingMode}, ${dataFormat}`;
-  it(testTitle, () => {
-    const x = tensor3d(xLength4Data, [1, 4, 1]);
-    let kernelData: number[] = [];
-    for (let i = 0; i < outChannels; ++i) {
-      kernelData = kernelData.concat(kernelLength2Data);
-    }
-    const kernel =
-      tfc.transpose(tensor3d(kernelData, [1, outChannels, 2]), [2, 0, 1]);
-    const y = K.conv1d(x, kernel, stride, paddingMode, dataFormat);
-    expectTensorsClose(y, tensor3d([-10, -10, -40, -40], [1, 2, 2]));
-  });
-});
-
-describeMathCPUAndGPU('conv2d', () => {
-  const x4by4Data = [[[
-    [10, 30, 50, 70], [20, 40, 60, 80], [-10, -30, -50, -70],
-    [-20, -40, -60, -80]
-  ]]];
-  const kernel2by2Data = [1, 0, 0, -1];
-
-  const dataFormats: DataFormat[] =
-    [undefined, 'channelsFirst', 'channelsLast'];
-  const paddingModes: PaddingMode[] = [undefined, 'same', 'valid'];
-  const stridesArray = [1, 2];
-
-  for (const dataFormat of dataFormats) {
-    for (const paddingMode of paddingModes) {
-      for (const stride of stridesArray) {
-        const testTitle = `stride=${stride}, ${paddingMode}, ` +
-          `${dataFormat}`;
-        it(testTitle, () => {
-          let x: Tensor = tensor4d(x4by4Data, [1, 1, 4, 4]);
-          if (dataFormat !== 'channelsFirst') {
-            x = tfc.transpose(x, [0, 2, 3, 1]);  // NCHW -> NHWC.
-          }
-          const kernel = tensor4d(kernel2by2Data, [2, 2, 1, 1]);
-          const y = K.conv2d(x, kernel, [stride, stride], 'valid', dataFormat);
-
-          let yExpected: Tensor;
-          if (stride === 1) {
-            yExpected = tensor4d(
-              [[[[-30, -30, -30], [50, 90, 130], [30, 30, 30]]]],
-              [1, 1, 3, 3]);
-          } else if (stride === 2) {
-            yExpected = tensor4d([[[[-30, -30], [30, 30]]]], [1, 1, 2, 2]);
-          }
-          if (dataFormat !== 'channelsFirst') {
-            yExpected = tfc.transpose(yExpected, [0, 2, 3, 1]);
-          }
-          expectTensorsClose(y, yExpected);
-        });
-      }
-    }
-  }
-});
-
-describeMathCPUAndGPU('conv2dWithBias', () => {
-  const x4by4Data = [[[
-    [10, 30, 50, 70], [20, 40, 60, 80], [-10, -30, -50, -70],
-    [-20, -40, -60, -80]
-  ]]];
-  const kernel2by2Data = [1, 0, 0, -1];
-  const biasScalarData = [2.2];
-
-  const outChannelsArray = [2, 3];
-  const dataFormats: DataFormat[] =
-    [undefined, 'channelsFirst', 'channelsLast'];
-  const paddingModes: PaddingMode[] = [undefined, 'same', 'valid'];
-  const stridesArray = [1, 2];
-
-  for (const outChannels of outChannelsArray) {
-    for (const dataFormat of dataFormats) {
-      for (const paddingMode of paddingModes) {
-        for (const stride of stridesArray) {
-          const testTitle = `outChannels=${outChannels}, stride=${stride}, ` +
-            `${paddingMode}, ${dataFormat}`;
-          it(testTitle, () => {
-            let x: Tensor = tensor4d(x4by4Data, [1, 1, 4, 4]);
-            if (dataFormat !== 'channelsFirst') {
-              x = tfc.transpose(x, [0, 2, 3, 1]);  // NCHW -> NHWC.
-            }
-
-            let kernelData: number[] = [];
-            let biasData: number[] = [];
-            for (let i = 0; i < outChannels; ++i) {
-              kernelData = kernelData.concat(kernel2by2Data);
-              biasData = biasData.concat(biasScalarData);
-            }
-            const kernel = tfc.transpose(
-              tensor4d(kernelData, [outChannels, 2, 2, 1]), [1, 2, 3, 0]);
-            const bias = tensor1d(biasData);
-
-            const y = K.conv2dWithBias(
-              x, kernel, bias, [stride, stride], 'valid', dataFormat);
-
-            let yExpectedShape: [number, number, number, number];
-            let yExpectedDataPerChannel: number[];
-            if (stride === 1) {
-              yExpectedShape = [1, outChannels, 3, 3];
-              yExpectedDataPerChannel =
-                [-30, -30, -30, 50, 90, 130, 30, 30, 30];
-            } else if (stride === 2) {
-              yExpectedShape = [1, outChannels, 2, 2];
-              yExpectedDataPerChannel = [-30, -30, 30, 30];
-            }
-            for (let i = 0; i < yExpectedDataPerChannel.length; ++i) {
-              yExpectedDataPerChannel[i] += biasScalarData[0];
-            }
-            let yExpectedData: number[] = [];
-            for (let i = 0; i < outChannels; ++i) {
-              yExpectedData = yExpectedData.concat(yExpectedDataPerChannel);
-            }
-            let yExpected: Tensor = tensor4d(yExpectedData, yExpectedShape);
-            if (dataFormat !== 'channelsFirst') {
-              yExpected = tfc.transpose(yExpected, [0, 2, 3, 1]);
-            }
-            expectTensorsClose(y, yExpected);
-          });
-        }
-      }
-    }
-  }
-});
-
-describeMathCPUAndGPU('depthwiseConv2d', () => {
-  const x4by4Data = [[[
-    [10, 30, 50, 70], [20, 40, 60, 80], [-10, -30, -50, -70],
-    [-20, -40, -60, -80]
-  ]]];
-
-  const dataFormats: DataFormat[] =
-    [undefined, 'channelsFirst', 'channelsLast'];
-  const paddingModes: PaddingMode[] = [undefined, 'same', 'valid'];
-  const stridesArray = [1, 2];
-  const depthMultipliers = [1, 2];
-
-  for (const dataFormat of dataFormats) {
-    for (const paddingMode of paddingModes) {
-      for (const stride of stridesArray) {
-        for (const depthMultiplier of depthMultipliers) {
-          const testTitle = `stride=${stride}, ${paddingMode}, ` +
-            `${dataFormat}, depthMultiplier=${depthMultiplier}`;
-          it(testTitle, () => {
-            let x: Tensor = tensor4d(x4by4Data, [1, 1, 4, 4]);
-            if (dataFormat !== 'channelsFirst') {
-              x = tfc.transpose(x, [0, 2, 3, 1]);  // NCHW -> NHWC.
-            }
-
-            let kernel: Tensor4D;
-            if (depthMultiplier === 1) {
-              kernel = tensor4d([1, 0, 0, -1], [2, 2, 1, 1]);
-            } else if (depthMultiplier === 2) {
-              // Two kernels of the same absolute values but opposite signs:
-              //   [[1, 0], [0, -1]] and [[-1, 0], [0, 1]].
-              kernel = tensor4d([1, -1, 0, 0, 0, 0, -1, 1], [2, 2, 1, 2]);
-            }
-            const y = K.depthwiseConv2d(
-              x, kernel, [stride, stride], 'valid', dataFormat);
-
-            let yExpected: Tensor;
-            if (stride === 1) {
-              if (depthMultiplier === 1) {
-                yExpected = tensor4d(
-                  [[[[-30, -30, -30], [50, 90, 130], [30, 30, 30]]]],
-                  [1, 1, 3, 3]);
-              } else if (depthMultiplier === 2) {
-                yExpected = tensor4d(
-                  [[
-                    [[-30, -30, -30], [50, 90, 130], [30, 30, 30]],
-                    [[30, 30, 30], [-50, -90, -130], [-30, -30, -30]]
-                  ]],
-                  [1, 2, 3, 3]);
-              }
-            } else if (stride === 2) {
-              if (depthMultiplier === 1) {
-                yExpected = tensor4d([[[[-30, -30], [30, 30]]]], [1, 1, 2, 2]);
-              } else if (depthMultiplier === 2) {
-                yExpected = tensor4d(
-                  [[[[-30, -30], [30, 30]], [[30, 30], [-30, -30]]]],
-                  [1, 2, 2, 2]);
-              }
-            }
-            if (dataFormat !== 'channelsFirst') {
-              yExpected = tfc.transpose(yExpected, [0, 2, 3, 1]);
-            }
-            expectTensorsClose(y, yExpected);
-          });
-        }
-      }
-    }
-  }
-
-  it('Non-4D kernel leads to exception', () => {
-    const x = tfc.zeros([1, 1, 4, 4]);
-    expect(() => K.depthwiseConv2d(x, tfc.zeros([1, 2, 2]), [
-      1, 1
-    ])).toThrowError(/.* is required to be 4-D, but is instead 3-D/);
-  });
-});
-
-describeMathCPUAndGPU('pool2d', () => {
-  const x4by4Data = [[[
-    [10, 30, 50, 70], [20, 40, 60, 80], [-10, -30, -50, -70],
-    [-20, -40, -60, -80]
-  ]]];
-  const x5by5Data = [[[
-    [0, 1, 3, 5, 7], [0, 2, 4, 6, 8], [0, 0, 0, 0, 0], [0, -1, -3, -5, -7],
-    [0, -2, -4, -6, -8]
-  ]]];
-
-  const poolModes: PoolMode[] = [undefined, 'max', 'avg'];
-  const dataFormats: DataFormat[] =
-    [undefined, 'channelsFirst', 'channelsLast'];
-  const stridesArray = [1, 2];
-  for (const poolMode of poolModes) {
-    for (const dataFormat of dataFormats) {
-      for (const stride of stridesArray) {
-        const testTitle = `4x4, ${stride}, same, ${dataFormat}, ` +
-          `${poolMode}`;
-        it(testTitle, () => {
-          let x: Tensor = tensor4d(x4by4Data, [1, 1, 4, 4]);
-          if (dataFormat !== 'channelsFirst') {
-            x = tfc.transpose(x, [0, 2, 3, 1]);  // NCHW -> NHWC.
-          }
-          let yExpected: Tensor;
-          if (poolMode === 'avg') {
-            if (stride === 1) {
-              yExpected = tensor4d(
-                [[[
-                  [25, 45, 65, 75], [5, 5, 5, 5], [-25, -45, -65, -75],
-                  [-30, -50, -70, -80]
-                ]]],
-                [1, 1, 4, 4]);
-            } else {
-              yExpected = tensor4d([[[[25, 65], [-25, -65]]]], [1, 1, 2, 2]);
-            }
-          } else {
-            if (stride === 1) {
-              yExpected = tensor4d(
-                [[[
-                  [40, 60, 80, 80], [40, 60, 80, 80], [-10, -30, -50, -70],
-                  [-20, -40, -60, -80]
-                ]]],
-                [1, 1, 4, 4]);
-            } else if (stride === 2) {
-              yExpected = tensor4d([[[[40, 80], [-10, -50]]]], [1, 1, 2, 2]);
-            }
-          }
-          if (dataFormat !== 'channelsFirst') {
-            yExpected = tfc.transpose(yExpected, [0, 2, 3, 1]);
-          }
-          const y = K.pool2d(
-            x, [2, 2], [stride, stride], 'same', dataFormat, poolMode);
-          expectTensorsClose(y, yExpected);
-        });
-      }
-    }
-  }
-
-  for (const poolMode of poolModes) {
-    it(`5x5, 2, same, CHANNEL_FIRST, ${poolMode}`, () => {
-      const x5by5 = tensor4d(x5by5Data, [1, 1, 5, 5]);
-      let yExpected = tensor4d(x4by4Data, [1, 1, 4, 4]);
-      if (poolMode === 'avg') {
-        yExpected = tensor4d(
-          [[[[0.75, 4.5, 7.5], [-0.25, -2, -3.5], [-1, -5, -8]]]],
-          [1, 1, 3, 3]);
-      } else {
-        yExpected =
-          tensor4d([[[[2, 6, 8], [0, 0, 0], [0, -4, -8]]]], [1, 1, 3, 3]);
-      }
-      const y =
-        K.pool2d(x5by5, [2, 2], [2, 2], 'same', 'channelsFirst', poolMode);
-      expectTensorsClose(y, yExpected);
-    });
-  }
-
-  for (const poolMode of poolModes) {
-    it(`5x5, 2, valid, CHANNEL_LAST, ${poolMode}`, () => {
-      const x5by5 =
-        tfc.transpose(tensor4d(x5by5Data, [1, 1, 5, 5]), [0, 2, 3, 1]);
-      let yExpected: Tensor4D;
-      if (poolMode === 'avg') {
-        yExpected = tensor4d([[[[0.75, 4.5], [-0.25, -2]]]], [1, 1, 2, 2]);
-      } else {
-        yExpected = tensor4d([[[[2, 6], [0, 0]]]], [1, 1, 2, 2]);
-      }
-      const y =
-        K.pool2d(x5by5, [2, 2], [2, 2], 'valid', 'channelsLast', poolMode);
-      expectTensorsClose(y, tfc.transpose(yExpected, [0, 2, 3, 1]));
-    });
-  }
-});
-
 describe('floatx ', () => {
   it('returns "float32"', () => {
     expect(K.floatx()).toEqual('float32');
@@ -1960,157 +1537,6 @@ describe('inTrainPhase', () => {
   });
   it('training = default false', () => {
     expect(K.inTrainPhase(() => -5, () => 5)).toEqual(5);
-  });
-});
-
-/**
- * A simplistic RNN step function for testing.
- * This step function simply
- * - calculates a reduced mean over all input elements, for each sample.
- * - adds that mean to the state tensor(s),
- * - take the negative of the 1st current state tensor and use it as the
- *   output.
- * @param inputs
- * @param states
- */
-function rnnStepForTest(inputs: Tensor, states: Tensor[]): [Tensor, Tensor[]] {
-  const mean = K.mean(inputs) as Scalar;
-  const newStates = states.map(state => K.scalarPlusArray(mean, state));
-  const output = tfc.neg(newStates[0]);
-  return [output, newStates];
-}
-
-describeMathCPUAndGPU('rnn', () => {
-  it('Simple step function: 3D inputs, 1 state', () => {
-    const inputs = tensor3d(
-      [[[1, 2], [3, 4], [5, 6]], [[10, 20], [30, 40], [50, 60]]], [2, 3, 2]);
-    const initialStates = [tfc.zeros([2, 4])];
-    const rnnOutputs = K.rnn(rnnStepForTest, inputs, initialStates);
-    const lastOutput = rnnOutputs[0];
-    const outputs = rnnOutputs[1];
-    const newStates = rnnOutputs[2];
-    expectTensorsClose(
-      lastOutput,
-      tensor2d(
-        [
-          [-57.75, -57.75, -57.75, -57.75],
-          [-57.75, -57.75, -57.75, -57.75]
-        ],
-        [2, 4]));
-    expectTensorsClose(
-      outputs,
-      tensor3d(
-        [
-          [
-            [-8.25, -8.25, -8.25, -8.25], [-27.5, -27.5, -27.5, -27.5],
-            [-57.75, -57.75, -57.75, -57.75]
-          ],
-          [
-            [-8.25, -8.25, -8.25, -8.25], [-27.5, -27.5, -27.5, -27.5],
-            [-57.75, -57.75, -57.75, -57.75]
-          ]
-        ],
-        [2, 3, 4]));
-    expect(newStates.length).toEqual(1);
-    expectTensorsClose(
-      newStates[0],
-      tensor2d(
-        [[57.75, 57.75, 57.75, 57.75], [57.75, 57.75, 57.75, 57.75]],
-        [2, 4]));
-  });
-
-  it('Simple step function: 3D inputs, 2 states', () => {
-    const inputs = tensor3d(
-      [[[1, 2], [3, 4], [5, 6]], [[10, 20], [30, 40], [50, 60]]], [2, 3, 2]);
-    // The two state tensors have different shapes.
-    const initialStates = [tfc.zeros([2, 4]), tfc.ones([2, 3])];
-    const rnnOutputs = K.rnn(rnnStepForTest, inputs, initialStates);
-    const lastOutput = rnnOutputs[0];
-    const outputs = rnnOutputs[1];
-    const newStates = rnnOutputs[2];
-    expectTensorsClose(
-      lastOutput,
-      tensor2d(
-        [
-          [-57.75, -57.75, -57.75, -57.75],
-          [-57.75, -57.75, -57.75, -57.75]
-        ],
-        [2, 4]));
-    expectTensorsClose(
-      outputs,
-      tensor3d(
-        [
-          [
-            [-8.25, -8.25, -8.25, -8.25], [-27.5, -27.5, -27.5, -27.5],
-            [-57.75, -57.75, -57.75, -57.75]
-          ],
-          [
-            [-8.25, -8.25, -8.25, -8.25], [-27.5, -27.5, -27.5, -27.5],
-            [-57.75, -57.75, -57.75, -57.75]
-          ]
-        ],
-        [2, 3, 4]));
-    expect(newStates.length).toEqual(2);
-    expectTensorsClose(
-      newStates[0],
-      tensor2d(
-        [[57.75, 57.75, 57.75, 57.75], [57.75, 57.75, 57.75, 57.75]],
-        [2, 4]));
-    expectTensorsClose(
-      newStates[1],
-      tensor2d([[58.75, 58.75, 58.75], [58.75, 58.75, 58.75]], [2, 3]));
-  });
-
-  it('Simple step function: 4D inputs, 2 states', () => {
-    const inputs = tensor4d(
-      [
-        [[[1], [2]], [[3], [4]], [[5], [6]]],
-        [[[10], [20]], [[30], [40]], [[50], [60]]]
-      ],
-      [2, 3, 2, 1]);
-    // The two state tensors have different shapes.
-    const initialStates = [tfc.zeros([2, 4]), tfc.ones([2, 3])];
-    const rnnOutputs = K.rnn(rnnStepForTest, inputs, initialStates);
-    const lastOutput = rnnOutputs[0];
-    const outputs = rnnOutputs[1];
-    const newStates = rnnOutputs[2];
-    expectTensorsClose(
-      lastOutput,
-      tensor2d(
-        [
-          [-57.75, -57.75, -57.75, -57.75],
-          [-57.75, -57.75, -57.75, -57.75]
-        ],
-        [2, 4]));
-    expectTensorsClose(
-      outputs,
-      tensor3d(
-        [
-          [
-            [-8.25, -8.25, -8.25, -8.25], [-27.5, -27.5, -27.5, -27.5],
-            [-57.75, -57.75, -57.75, -57.75]
-          ],
-          [
-            [-8.25, -8.25, -8.25, -8.25], [-27.5, -27.5, -27.5, -27.5],
-            [-57.75, -57.75, -57.75, -57.75]
-          ]
-        ],
-        [2, 3, 4]));
-    expect(newStates.length).toEqual(2);
-    expectTensorsClose(
-      newStates[0],
-      tensor2d(
-        [[57.75, 57.75, 57.75, 57.75], [57.75, 57.75, 57.75, 57.75]],
-        [2, 4]));
-    expectTensorsClose(
-      newStates[1],
-      tensor2d([[58.75, 58.75, 58.75], [58.75, 58.75, 58.75]], [2, 3]));
-  });
-
-  it('Using inputs <3D leads to ValueError', () => {
-    const inputs = tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const initialStates = [tfc.zeros([4]), tfc.ones([3])];
-    expect(() => K.rnn(rnnStepForTest, inputs, initialStates)).toThrowError();
   });
 });
