+++ src/layers/padding.ts
@@ -16,7 +16,7 @@
 //   but we decided to put them in a separate file (padding.ts) for clarity.
 
 // tslint:disable:max-line-length
-import {serialization, Tensor} from '@tensorflow/tfjs-core';
+import {pad, serialization, Tensor} from '@tensorflow/tfjs-core';
 
 import {imageDataFormat} from '../backend/common';
 import * as K from '../backend/tfjs_backend';
@@ -27,6 +27,82 @@ import {Kwargs, Shape} from '../types';
 import {getExactlyOneShape, getExactlyOneTensor} from '../utils/generic_utils';
 
 // tslint:enable:max-line-length
+
+/**
+ * Pads the middle dimension of a 3D tensor.
+ *
+ * @param x Input `Tensor` to be padded.
+ * @param padding `Array` of 2 integers, how many zeros to add at the start and
+ *   end of the middle dimension (i.e., dimension 1).
+ * @return A padded 3D `Tensor`.
+ */
+export function temporalPadding(x: Tensor, padding?: [number, number]): Tensor {
+  if (K.ndim(x) !== 3) {
+    throw new ValueError(
+        `temporalPadding expects input tensor to be 3-D, but received a ` +
+        `${K.ndim(x)}-D tensor.`);
+  }
+
+  if (padding == null) {
+    padding = [1, 1];
+  }
+  if (padding.length !== 2) {
+    throw new ValueError(
+        `temporalPadding expects input padding pattern to be a length-2 ` +
+        `array, but received a length-${padding.length} array.`);
+  }
+
+  const pattern: Array<[number, number]> = [[0, 0], padding, [0, 0]];
+  return pad(x, pattern);
+}
+
+/**
+ * Pads the 2nd and 3rd dimensions of a 4D tensor.
+ *
+ * @param x Input `Tensor` to be padded.
+ * @param padding `Array` of two `Array`s, each of which is an `Array` of two
+ *   integers. The amount of padding at the beginning and end of the 2nd and 3rd
+ *   dimensions, respectively.
+ * @param dataFormat 'channelsLast' (default) or 'channelsFirst'.
+ * @return Padded 4D `Tensor`.
+ */
+export function spatial2dPadding(
+    x: Tensor, padding?: [[number, number], [number, number]],
+    dataFormat?: DataFormat): Tensor {
+  if (K.ndim(x) !== 4) {
+    throw new ValueError(
+        `temporalPadding expects input tensor to be 4-D, but received a ` +
+        `${K.ndim(x)}-D tensor.`);
+  }
+
+  if (padding == null) {
+    padding = [[1, 1], [1, 1]];
+  }
+  if (padding.length !== 2 || padding[0].length !== 2 ||
+      padding[1].length !== 2) {
+    throw new ValueError(
+        'spatial2dPadding expects `padding` to be an Array of two Arrays, ' +
+        'each of which is an Array of two integers.');
+  }
+
+  if (dataFormat == null) {
+    dataFormat = imageDataFormat();
+  }
+  if (dataFormat !== 'channelsLast' && dataFormat !== 'channelsFirst') {
+    throw new ValueError(
+        `Unknown data format: ${dataFormat}. ` +
+        `Supported data formats are 'channelsLast' and 'channelsFirst.`);
+  }
+
+  let pattern: Array<[number, number]>;
+  if (dataFormat === 'channelsFirst') {
+    pattern = [[0, 0], [0, 0], padding[0], padding[1]];
+  } else {
+    pattern = [[0, 0], padding[0], padding[1], [0, 0]];
+  }
+
+  return pad(x, pattern);
+}
 
 export interface ZeroPadding2DLayerConfig extends LayerConfig {
   /**
@@ -164,7 +240,7 @@ export class ZeroPadding2D extends Layer {
   }
 
   call(inputs: Tensor|Tensor[], kwargs: Kwargs): Tensor|Tensor[] {
-    return K.spatial2dPadding(
+    return spatial2dPadding(
         getExactlyOneTensor(inputs), this.padding, this.dataFormat);
   }
